Remove special rule for debian make clean
This special rule removed the build directory, which is previously
created by the jenkins job and is where the build takes place.
The build must take place in the directory named "build"
in order for the artifacts to be located in the proper path
for the install_eos.sh script.

[endlessm/eos-shell#155]