Adjusted the ramp thresholds to finish fading out the logo at 100%
Also, made the logo and the spinner stay in 'full' state between 45% and
60% before ramping down, so the % of time the logo will be fading out
now is not that different than before (it faded out for 18% of the time).

Note that the time the logo would be fading out in an ideal case would
be now 25%, which is more than before, but I think it's probably a
better idea to leave it like this than start ramping down at a later
point so the logo would be almost invisible by the end of the booting
process, when it's still likely that the animation would get stuck for a
few seconds (yet it's probably better to see that than a white screen).

[endlessm/eos-shell#3277]